[SH] Workaround for SMI bit #15 issue on GMAC#0 on B2000 (STxH415) Board
This commit adds a (short-term) workaround for an issue that seems to
affect the SMI (Serial Management Interface) on GMAC #0 on the B2000
(STxH415) Validation Board. The issue is that when reading the 16-bit
PHY registers, bit #15 "randomly" appears to be erroneously read as a
'1', when it should be a '0'.

This problem has not (yet) been fully analysed, and so this workaround
is provided "as is", and should ideally be removed once the root cause
is properly analysed and understood.

So far, it appears that it is *only* bit #15 that is impacted, and it
is only false '1's that have been observed.  No corruptions have been
noticed on GMAC #1.  Also, it looks like it is only reads that are
affected. This problem has been noticed on several different B2000B
boards, and the amount of "randomness" varies from board to board.

This workaround may prove to be insufficient, if bit #15 is *always*
wrong, and not just "mostly" wrong! So, it may not work on all boards.

When a potentially erroneous bit is encountered, the STMAC driver will
"clear" it to a '0', and will generate an appropriate diagnostic.

All testing was on a single board, "B2000B-00 N026", mated with a
"B2032B-00 N256" daughter-board on CN22 (GMII0).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>