SPEAr13xx: Disable NAND write verification
CONFIG_MTD_NAND_VERIFY_WRITE enables the NAND write verify routine. This routine
compares the content written to NAND device bit by bit with the source content.

With the advent of high capacity and high density devices, the probability of
getting a bit error increases. In such cases, a nand_verify_buf routine returns
a failure leading to NAND writes getting aborted abruptly.

Since spear FSMC supports correction of up to 8 bits in 512 bytes, this verify
routine is inefficient since it reports errors even in cases when the errors are
correctable. There are two solutions to the problem
- Read the page using ecc correction even in verify_buf routines
- Disable nand write verification

This patch implements the second option for spear13xx devices

Signed-off-by: Vipin Kumar <vipin.kumar@st.com>