[SH] Adds initial support for RMII (B2035) on B2000 (STxH415) Validation Board
This commit adds initial support for RMII mode using the B2035A (IC+
IP101A) daughter-board on STMicroelectronics' B2000 Validation Board,
which uses STMicroelectronics' STxH415 (Orly) SoC device.

However, it should be noted that at time of the commit, RMII mode on the
B2035A was *not* working! The code appears to be complete and correct,
and has been independently reviewed. It should be noted that the same
H/W successfully works with ARM linux, but consistently fails with ST40
U-Boot. Further analysis is clearly required, but it is worthwhile
committing this change "as is", as it should be close to functional!
Hopefully, it should just a one-line change that is required...

All testing was on a single board, "B2000B-00 N026", using a B2035A-00
N011 daughter-board.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>