[SH] Initial support for the B2057 (STxH207) Reference Board
This commit adds initial support for the B2057 (STiH207-HDK) Reference
Board, which uses the STMicroelectronics' STiH207 (Liege) SoC device.

The following features have now been successfully tested:

	* 32-bit (SE) mode
	* Serial ASCs (2 off)
	* SPI Serial Flash (FSM and "bit-banging")
	* Boot from SPI Serial Flash
	* I2C bus

The following features are to be completed:

	* 4-bit ECC SLC NAND Flash (using BCH controller)
	* Boot from NAND
	* Ethernet
	* USB (OHCI)
	* SATA

The following features are not planned to be supported:

	* 29-bit mode
	* NOR flash (no NOR on the board)
	* PCIe

Note:
	The SPI Serial FLASH device fitted has a capacity of 32MiB,
	however, U-Boot still only uses 24-bit addresses for SPI,
	hence only the first 16MiB of the SPI device is usable. Hence
	the device is effectively capped to the first 16MiB only.

All testing was on a single B2057A board, "STiH207-HDKAST N001".

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>