net/macb: workaround for transmission hang issue
It is observed on SPEAr320S RMII#1 interface that on transmitting
packets the MAC dma hangs randomly and constantly showing busy tx-go
state.

It comes out if this situation only when Transmission is disabled and
enabled again.

Since it happens randomly and u-boot doesn't require high performance we
disable TE and re-enable it on each transmission. We also change number
of transmit descriptor to 1 as we would not require more than it, further
it would not alter GMAC notion of transmit descriptor start queue as it
always point to same descriptor.

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>