[ST40] Add missing ST40_PIO_BASE(x) macros for ST40 SoC headers
This commit just adds the macro "ST40_PIO_BASE(x)" to the header files
"include/asm-st40/<soc>reg.h" which were missing this macro.

This macro just expands out to one of the many ST40_PIO<x>_REGS_BASE
macros, which yields the base address of PIO bank #x. Having a single
ST40_PIO_BASE(x) macro for this purpose is very useful, and most recent
ST40 boards benefit from having this macro. This commit just adds it to
the older boards, when it was not ubiquitous.

Note: This macro should evaluate just to a simple pointer/address as long
as "x" is a integer literal, or a simple expression which can be evaluated
at pre-processor-time or compile-time, hence it may be compile-time bloat,
but it should (always?) be run-time efficient.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>