[ST40] Complete Merge of 'v1.3.4' into STMicroelectronics' ST40 tree
This commit completes the merge of the official 'v1.3.4' release (which
includes all the currently supported STMicroelectronics' ST40 boards.

The main thing that the previous few commits failed to address, was
that of "testing".  This commit confirms that some (albeit limited)
testing has now successfully taken place on (just two) ST40 boards,
against the recently merged 'v1.3.4' tree.  Testing explicitly included
the following aspects (each nominally on just one board):

	I2C, USB (OHCI), Ethernet (with "ping")
	NOR, NAND and SPI Flashes
	SPI with both "bit-banging", and FSM drivers
	boot-from-NOR, boot-from-SPI,
	booting ST40 linux (with "bootm") from (parallel) NOR flash

Testing of the above was successful with no regressions, and only a
couple of very minor issues (just print formatting) were observed, whose
fixes are included herein. Testing focused mainly on SPI, as that was the
sub-system that probably had the greatest amount of manual merging work,
and the greatest perceived risk.  More testing (on different boards+SoCs)
clearly needs to be undertaken, but to a first order of approximation,
there does not appear to be any fundamental problems with the recent
re-baselining to 'v1.3.4', at least for for ST40 boards.

All the above testing was performed on just two boards:
	1) HDK-7108 board (STi7108-HDK10ZA-N206)
	2) B2057A   board (STiH207-HDKAST N001)

Specifically, *all* the delinquencies are now (believed to be) addressed,
and there are no (known) outstanding issues with the merge to 'v1.3.4'.

	see commit e2f47d9fdeec563b611e336e1f06dda4635ceb42

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>