[ST40] Rename CFG_ENV_xxxx to CONFIG_ENV_xxxx for the ST40 tree
The previous commit (1479df7f15dbb7bde5b3193b7fd4f388e790a284) was a
"snapshot", which did not build cleanly. This commit is an augmentation
of that commit, en route a fully working merge of 'v2008.10'.

This commit just renames several configuration macro names which were
originally included as part of the 'v1.3.4' to 'v2008.10' set of commits
on the master branch. The macros in question are:

	CFG_ENV_xxxx         -> CONFIG_ENV_xxxx
	CFG_FLASH_CFI_DRIVER -> CONFIG_FLASH_CFI_DRIVER

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>