[ST40] Initial support for nand_chip.hwcontrol() -> nand_chip.cmd_ctrl()
A recent commit (1479df7f15dbb7bde5b3193b7fd4f388e790a284) was a
"snapshot", which did not build cleanly. This commit is an augmentation
of that commit, en route to a fully working merge of 'v2008.10'.

This commit just changes both the prototype, definition, and all the
callers of the ST40-specific function stm_default_board_nand_init(). This
is required, as one of its parameters is a function pointer, defined by
the field "hwcontrol" field in "struct nand_chip", which has changed both
the number of arguments, the field has also been renamed to "cmd_ctrl",
as part of the 'v1.3.4' to 'v2008.10' set of commits on the master branch.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>