[ST40] fix do_bootm_linux() in lib_st40/bootm.c for ST40 merge
The recent commit (1479df7f15dbb7bde5b3193b7fd4f388e790a284) was a
"snapshot", which did not build cleanly. This commit is an augmentation
of that commit, en route a fully working merge of 'v2008.10'.

This commit essentially just deals with modifications required to be
made to the do_bootm_linux() function in lib_st40/bootm.c, as a result
of the image handling infrastructure that changed between the 'v1.3.4'
and the 'v2008.10' tagged releases.  The file now compiles cleanly.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>