[NAND] Additional diagnostic when we have probed an unrecognised NAND device
This commit is a partial recapitulation (improvement, and
contemporaneous version) of an earlier STMicroelectronics-specific commit
0ed67c53054ed1e3deeda749747eab1c44926270.  Please see the original commit
for further details.

This commit just prints out the Manufacturer's ID and the Device ID when
an "unknown" NAND device is probed. This is useful when we are starting
to see new boards with new NAND devices on them, but U-Boot does not
know what they are.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>