[ST40] Complete Merge of 'v2008.10' into STMicroelectronics' ST40 tree
This commit completes the merge of the official 'v2008.10' release (which
includes all the currently supported STMicroelectronics' ST40 boards.

Testing, so far, was performed on just one board - the HDK-7108 board
(STi7108-HDK10ZA-N206). Testing focused on NAND, as that was the area of
greatest change.  More testing (on different boards+SoCs) clearly needs
to be undertaken, but to a first order of approximation, there does not
appear to be any fundamental problems with the recent re-baselining to
'v2008.10', at least for one ST40 board.

Specifically, *all* the delinquencies are now (believed to be) addressed,
and there are no (known) outstanding issues with the merge to 'v2008.10'.

	see commit 1479df7f15dbb7bde5b3193b7fd4f388e790a284

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>