[FIX] Allow ONFI to probe for NAND devices not in nand_flash_ids[]
Prior to this commit, the pointer "type" (in nand_get_flash_type) could
be NULL, if the NAND device's ID was not found in the nand_flash_ids[]
array of known NAND devices. This could result in NULL pointer
dereferences. Also, the code would end up returning prematurely, *before*
it even probed for a ONFI compliant NAND device.

Now, if we fail to find a match in nand_flash_ids[], "type" will end
up pointing to the *last* element in the array, which is initialized to
all zeros, which is safe.

In addition, we will now always probe for ONFI-compliant devices first,
even if the probed NAND is not located in nand_flash_ids[]. Finally, we
now issue a "Unknown NAND" diagnostic only if it is not ONFI-compliant,
and the probed NAND is not found in the nand_flash_ids[] array.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>