[ST40] Rename #defines for ST's NAND HAMMING controller
This commit just renames all the #define macros pertaining to
STMicroelectronics EMI NAND "Hamming" controller, to include "_HAM_"
in their names. That is:

	rename ST40_EMI_NAND_xxxx -> ST40_EMI_NAND_HAM_xxxx

This is a pre-cursor to the addition of support for STMicroelectronics
EMI NAND "BCH" controller, which uses similar register names, but the
two occlude each other, and having the two sets of #defines in different
names-spaces is considered to be cleaner. The new (not yet added) names
for ST's BCH controller will be of the form: ST40_EMI_NAND_BCH_xxxx.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>