[ST40] re-map BCH's "nand dump", logically by separating DATA and ECC areas
The BCH controller in STMicroelectronics' NANDi block, stores the DATA
and ECC in interleaved "stripes" (1KiB data, ECC, 1KiB data, ECC, ...)
Hence the in-band region includes both DATA and ECC, and the same is
also true for the OOB region as well.

This commit allows "nand dump" optionally to re-map the DATA and ECC
parts, such that we move all the DATA bytes logically to be all in-band,
and similarly, we move all the ECC bytes logically to be all OOB.
This helps intuitively to see what is going on, but we are really lying
to the user about what is happening physically inside the NAND device. To
see the *raw* physical layout, then simply set the environment variable
"bch_raw", before running "nand dump".

This is only enabled if the macro CFG_ST40_NAND_USE_BCH is defined.
If such a macro is defined, then this feature is enabled by default,
and can be disabled dynamically by simply just setting the environment
variable "bch_raw" (to anything).

Also, as a special case, we do no re-mapping at all, if eec.bytes == 0,
i.e. if NO ECC is used by the BCH -- not recommended!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>