[NAND] Improve ONFI to obtain a better "description" of the NAND device
In the raw ONFI structures, the NAND device "model" is stored without
NUL-termination, and (normally) also with trailing white space. Using
it like this, as a classic "C string", is best avoided.

This commit will now properly NUL terminate the raw ONFI "model" string,
and trim all trailing white spaces from the original string. In addition,
we add both the total size and the page size information to the "name"
string stored in the "mtd" structures, which can now be easily seen by
running the "nand info" command.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>