[ST40] Adds initial support for multi-bit BCH support on ST40
Adds initial support for multi-bit (off-die) ECC SLC NAND devices, using
STMicroelectronics' BCH controller in the NANDi IP block.  By "off-die",
I mean the ECC computation is *not* performed inside the NAND package,
but on the embedded/host system, e.g. by a STxH205 SoC.

STMicroelectronics' BCH controller can detect and correct up to a maximum
of either 18-bits, or 30-bits, (both per 1024 byte sector). ECC selection
depends on the exact size of the OOB region in the SLC NAND device that
is being used - specifically:

	*) 18-bits of correction requires 32 bytes of OOB per 1KiB
	   of data (e.g. at least 128 bytes of OOB on a 4KiB page).

	*) 30-bits of correction requires 54 bytes of OOB per 1KiB
	   of data (e.g. at least 216 bytes of OOB on a 4KiB page).

This commit adds support for both the 18-bit and 30-bit ECC (per 1KiB
sector) modes of operation that is supported on STMicroelectronics'
BCH controller. To enable this support then the following macro should
be defined:

	#define CFG_ST40_NAND_USE_BCH

In addition, one (and only one) of the following macros must be defined
to specify the ECC "potency" to be used:

	#define CFG_ST40_NAND_USE_BCH_NO_ECC
	#define CFG_ST40_NAND_USE_BCH_18_BIT_ECC
	#define CFG_ST40_NAND_USE_BCH_30_BIT_ECC

It is the user's responsibility to ensure that U-Boot and linux are
both configured to use the same ECC potency. Similarly, for booting from
multi-bit NAND devices, the user has to ensure the "mode-pins" are set
to agree with the ECC potency that U-Boot + linux uses. Failure to do
this will result in undefined behaviour, and most likely the contents of
the BBTs being irretrievably lost, which could be a significant problem,
as the original manufacturers' bad block markers will be over-written
when the block is first erased,, and only stored in the BBTs.

So far, reading and writing with both 18-bit and 30-bit ECC appears
to work correctly. In addition, the generation and recognition of
NAND-resident Bad Block Tables (BBTs) also appears to work correctly.
However, testing has been limited to only a single board: B2057A board
(STiH207-HDKAST N001), With a Micron MT29F8G08ABABAWP, a 1024 MiB SLC
NAND with a page size of 4096+224, and sector size 512 KiB.

Note, this is an *initial* commit, and further testing of this is
required, as there have been a significant amount of changes!

Of particular interest, it should be noted that the entirety of the OOB
region is reserved for the H/W BCH ECC engine, and the S/W can not use
*any* of the OOB at all!  As a result the special "magic BBT signature"
has been forced from the traditional out-of-band area, to being "in-band".
The exact layout and placement scheme of this magic signature is liable
to change ...  (and is highly likely to change!!!).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>