[ST40] Make bch_wait_till_completion() safer (use SEQ_NODES_OVER)
Make the function bch_wait_till_completion() safer, by polling on the
SEQ_NODES_OVER interrupt status bit, rather than the RUN_STATUS in the
sequence status register.

This is required as the BCH engine completes *before* the DMA transfers
have completed (for page reads), hence if one used RUN_STATUS, then one
could try and read the freshly read data before the DMA had finished
writing it into RAM, which we must avoid doing!

Note: although we are using the interrupt status register, we are *not*
using interrupts, we will simply poll on the bit in the interrupt status
register to detect safe conclusion of the sequence.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>