[ST40] Added comment to justify read-after-write for NANDi switching
In emiss_nandi_select(), when we select a new NAND controller, we read
the configuration register back, after writing a new value to it. This
commit simply just adds a comment to explain that the need for this
was empirically required by Linux, on the Freeman Ultra system.

It is reasonable to assume that we should (for safety) also do this on
all ST40 SoCs on U-Boot.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>