[ST40] Add support for booting from multi-bit ECC using BCH controller
Added support for booting from NAND using STMicroelectronics'
BCH (multi-bot ECC) NAND controller.  Specifically, the following
boot-from-NAND modes with the BCH controller are now all supported,
and tested:

	Boot from 8-bit SLC NAND, 4KiB pages, 5 address cycles, with
	the (multi-bit ECC) BCH controller (not Hamming).
	There are three different possible BCH ECC modes:

		a) NO ECC (i.e. for "error free" NAND)
		b) 18-bits of ECC per 1KiB sector (needs 32 bytes OOB/1KiB)
		c) 30-bits of ECC per 1KiB sector (needs 54 bytes OOB/1KiB)

One (and only one) of the following macros were each used, in turn,
to build, to burn and to boot-from-NAND, to confirm correct behaviour.

	#define CFG_ST40_NAND_USE_BCH_NO_ECC
	#define CFG_ST40_NAND_USE_BCH_18_BIT_ECC
	#define CFG_ST40_NAND_USE_BCH_30_BIT_ECC

For each of the above (in turn), the NAND was "scrubbed," and a new BBT
(Bad Block Table) was written afresh. The "u-boot.bin" file was written
to physical block #0, and the environment variables was written (and read)
several times. The board was reset, and no issues were noted.

Testing has been limited to only a single board: B2057A board
(STiH207-HDKAST N001), With a Micron MT29F8G08ABABAWP, a 1024 MiB SLC
NAND with a page size of 4096+224, and sector size 512 KiB.

It should be noted that as the MT29F8G08ABABAWP has a (very large)
erase sector size of 512KiB, this means that "u-boot.bin" fits in
a single sector, and wastes most of that sector, as it is probably
undesirable to use the remainder of the boot sector for any real "data".
More interestingly, and using the same argument, is that the environment
variables (typically only a few KiB) typically will also need a dedicated
512KiB sector, which is very wasteful! This is not a real problem, it
just means that in practice that U-Boot (u-boot.bin + environment) now
will reserve 1MiB of the FLASH! Clearly devices with larger erase block
sizes, will still need 2 blocks just for U-Boot (and its environment).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>