[ST40] Add initial Networking support for B2057 (STxH207)
Added support for networking with in-package IC+ IP101A PHY, in MII mode.

Still to add support for RMII mode, and the IC+ IP101G PHY.

Note: it is required to ensure that the 3 jumpers to select MII/RMII mode
are properly configured on the B2057 board, see comment in board_init()
in the file "board/st/b2057/b2057.c" for further details.

Testing has been limited to only a single board: B2057A board
(STiH207-HDKAST N001), with a STxH207 SoC with an in-package IC+ IP101A
ethernet PHY.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>