[ST40] Add Networking support for IC+ IP101G on B2057 (STxH207)
Added support for networking with in-package IC+ IP101G PHY, in MII mode,
augmenting to the previous commit which added support only for the IP101A.

This update now provides support for both the IC+ IP101A and the IC+
IP101G ethernet PHYs, either of which may be in-package on the STxH207
SoC.

Note: Only MII mode is tested, still to add support for RMII mode.

Note: it is required to ensure that the 3 jumpers to select MII/RMII mode
are properly configured on the B2057 board, see comment in board_init()
in the file "board/st/b2057/b2057.c" for further details.

Testing has been limited to only a single board: B2057A board
(STiH207-HDKAST N047), with a pair of STxH207 SoCs, one with an in-package
IC+ IP101G ethernet PHY, and the other with an in-package IC+ IP101A
ethernet PHY. The same U-Boot binary image in FLASH works with both
varieties of PHYs, and was able to auto-detect which PHY is present.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>