[STMAC] Revert "Allow DMA to initialize when auto-negotiated to Gigabit"
This logically reverts commit b8fe5fbac4c2496fc462c85d3775937b35215376.

It transpires that although b8fe5fbac4c2496fc462c85d3775937b35215376
empirically did appear to fix the problem observed on the B2000 board
when connected to a Gigabit capable PHY, that fix (in hindsight) was a
"bad fix", as it later transpired that it broke MII on (at least) the
PDK-7105 board, and also broke RMII on (at least) the B2057 board. Hence,
this commit locally reverts the referenced commit.

Note this is not a "full" revert, as two of the '#if 0' ... '#endif'
diagnostics introduced in b8fe5fbac4c2496fc462c85d3775937b35215376
are still present after this reversion, and are potentially still
useful. Hence, it is a "logical" revert, of the active code, and not of
the commented-out diagnostics!

Note, as a result of this commit, the problem the original commit was
trying to fix was now been reactivated, and a "proper" solution for this
now needs to be sought!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>