[ST40] Add RMII Networking support for IC+ IP101x on B2057 (STxH207)
Added support for RMII to the existing MII support for networking with
the in-package IC+ IP101A/IP101G PHY on the B2057 board, which uses the
STxH207 SoC.

This update now provides support for both RMII and MII modes using both
the IC+ IP101A and the IC+ IP101G ethernet PHYs, either of which may be
in-package on the STxH207 SoC.

Note: it is required to ensure that the 3 jumpers to select MII/RMII mode
are properly configured on the B2057 board, see the comment in the file
"include/configs/b2057.h" for further details.

Also, to enable RMII mode, the macro CONFIG_STM_USE_RMII_MODE should be
defined in the file "include/configs/b2057.h", otherwise MII mode will
be implicitly used.

Testing has been limited to only a single board: B2057A board
(STiH207-HDKAST N047), with a pair of STxH207 SoCs, one with an in-package
IC+ IP101G ethernet PHY, and the other with an in-package IC+ IP101A
ethernet PHY.  Both RMII and MII modes were successfully tested.

Note: in both RMII and MII modes, only an external clock was tested.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>