imls: Reduce the heap requirement of imls
The imls implementation for NAND devices allocates an area equal to erasesize of
device. Since there are devices with > 2M erase size supported by the nand
driver framework, the implementation of imls needs an increase in MALLOC_LEN by
2M. This allocation area lies below the load address of the u-boot (0x700000). A
short memory map is as below

	--------- 0x700000
	|	|
	|  heap	|
	|	|
	--------- XXXXXXXX
	| stack	|
	|   |	|
	|   v	|
	---------

As shown in the map, the stack grows downwards and after the imls support for
NAND devices patch, the stack went down to 0x2f000000.

Another area of concern is the linux load address (0x8000) and ever increasing
linux kernel size. The above change leads to limit the kernel binary size that
can be booted with bootm command

This patch modifies the implementation of imls to allocate only writesize (nand
device page size) bytes to obviate the need to increase MALLOC_LEN.

Signed-off-by: Vipin Kumar <vipin.kumar@st.com>