spear1340: Add SSP Chip Select control option through software
The SSP cs can be controlled either by ssp controller or by controlling the cs
pad through hardware. When ssp controller controls the cs, there is a high
probability of FIFO getting flushed too early (with in the polling latency) and
the cs being released. This results in broken transfers.

The way out is to use software control of the ssp cs pad and hold it low until
the transfer is complete. This patch implements this logic.

Signed-off-by: Vipin Kumar <vipin.kumar@st.com>