pl022: Modify the pre-scaler calculation routine
pl022 ssp controller can generate a frequency to interact with the external SPI
devices which is based on the following equation

		PClk
SSP Clk = -----------------
	  CPSDVR * (1 +SCR)

Until now, the logic only uses CPSDVR to program the frequency and thus limiting
the pre-scaler to 254. This patch tries to use both SCR as well as CPSR to
program the pre-scaler.

Signed-off-by: Vipin Kumar <vipin.kumar@st.com>