[NAND] Use ONFI probed name in preference to one in nand_flash_ids[]
Prior to this commit, if an ONFI compliant NAND device existed in the
nand_flash_ids[] array, then the final "description" of it would be the
static one stored in nand_flash_ids[], and not the (better) one generated
in nand_flash_detect_onfi(), during ONFI probing.

This commit will now use the name obtained from the ONFI probing (if
found), in preference to the static name in the nand_flash_ids[] array.

Hence, "nand info", should now provide the name from ONFI, if possible,
even for devices also in nand_flash_ids[].  For ONFI devices missing
from nand_flash_ids[], then there should be no change in functionality,
as the ONFI name will still be used, as before.

This fix was applied after noticing the non-ONFI name emitted when
using the NAND08GW3B2CN6 device, which is both ONFI compliant, *and*
present in nand_flash_ids[].

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>