[ST40] Add initial support for NAND BCH on the HDK-7108 board.
This commit adds initial support for BCH NAND on STMicroelectronics'
STx7108 SoC device, on the HDK-7108 platform.

Please note that support for the BCH controller is only present in cut
2.x of the STx7108 SoC, that is, BCH is not possible in cut 1.x of the
STx7108 SoC.

It should be noted that BCH support is *still* disabled by default, and
the default mode is to use the (legacy) Hamming (1-bit ECC) controller.
To enable the BCH (multi-bit) controller, then the following needs to
be configured:

In addition, the appropriate BCH ECC "potency" (0-bit, 18-bit, 30-bit)
must be selected in "include/configs/hdk7108.h".

Only minimal testing of this support has been done, and only on a single
board (STi7108-HDK22-N814).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>