[NAND] Bug-fix for probing non-ONFI compliant NANDs, when ONFI enabled
This commit fixes a bug, whereby, a non-ONFI compliant NAND device
would be miss-probed, when CONFIG_SYS_NAND_ONFI_DETECTION is defined.
Essentially, the device ID as returned by the READID command could be
"shifted", and hence incorrectly decoded, resulting in the NAND device's
geometry being wrongly attributed.

Obviously, there is a simple workaround:
	#undef CONFIG_SYS_NAND_ONFI_DETECTION

This commit fixes this issue properly.

Analysis:
========

Prior to this commit, when CONFIG_SYS_NAND_ONFI_DETECTION is defined,
and the probed NAND device is *not* ONFI compliant, then the READID
with an address of 0x20, will be treated as a "normal" READID command
(address 0x00), and not the special "ONFI" READID command.

So, the (flattened) code sequence is broadly:

	chip->cmdfunc(mtd, NAND_CMD_RESET, -1, -1);
	chip->cmdfunc(mtd, NAND_CMD_READID, 0x00, -1);
	*maf_id = chip->read_byte(mtd);			/* read ID[0] */
	*dev_id = chip->read_byte(mtd);			/* read ID[1] */
	...
	chip->cmdfunc(mtd, NAND_CMD_READID, 0x20, -1);
	if (chip->read_byte(mtd) != 'O' || chip->read_byte(mtd) != 'N' ||
		chip->read_byte(mtd) != 'F' || chip->read_byte(mtd) != 'I')
		return 0;
	...
	chip->cellinfo = chip->read_byte(mtd);		/* read ID[2] */
	extid = chip->read_byte(mtd);			/* read ID[3] */

Hence, the read_byte(mtd) which we compare with 'O', is actually ID[0]
(same as *maf_id). The comparison will fail, so we stop looking for the
magic 'ONFI' signature.  Then the next read_byte(mtd), which we assign to
chip->cellinfo, is *actually* ID[1], and not ID[2] as we expect. Hence,
we have an "off-by-one" with the device ID, and bad things happen...

The solution is to read all (4) bytes of the device ID atomically
(without performing any additional READID commands).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>