[ST40] Import latest re-timing values for MII+RMII on the STx7108.
This commit imports the recent PIO "re-timing" changes in the function
stx7108_configure_ethernet() and the associated updated data structures in
drivers/stm/stx7108_hispeed.c from STLinux, based on the "stlinux24_0210"
tag for the STx7108 SoC.

Both MII and RMII modes have been tested (albeit only using GMAC #1).

To change between MII and MII modes, then the "mode" field should
be set to stx7108_ethernet_mode_mii or stx7108_ethernet_mode_rmii,
as appropriate in the stx7108_ethernet_config structure passed to
stx7108_configure_ethernet().

Note: When using MII, then the PHY clock should be set to 25MHz, but
it should be set to 50MHz when using RMII. When using an "internal"
clock (i.e. not an external clock), then it should also be noted that
the current U-Boot code does *not* modify any of the PHY clocks ... it
is expected that the TargetPacks (or some other mechanism) is used to
set the PHY clock to the correct frequency.

Testing was conducted on two boards - MII on a STi7108-HDK22ZB-E1022
board, and RMII on a customer board using a STx7108+LAN8720 combination.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>