[ST40] Set internal PHY clock to 50MHz for RMII on STx7108.
This commit will "adjust" the PHY clock frequency to 50MHz, when in RMII
mode, when an internal PHY clock is used on the STx7108.

Specifically, this code will half the PHY clock divider, hence, doubling
the PHY clock frequency. This is based on the key assumption that the
TargetPacks will configure the PHY clock to 25MHz (for MII mode), hence
we can double the frequency correctly to 50MHz for RMII mode.

WARNING: If the above assumption is false, and the PHY clock is *not*
running initially at 25MHz as we expect, then RMII mode will not work
correctly.  At time of writing, the TargetPacks sets both CLK_ETH_PHY_0
and CLK_ETH_PHY_1 to be 25MHz, to be compatible with MII mode.

This has been tested on a single customer board, using only GMAC #1.
Testing of RMII on a customer board was a STx7108+LAN8720 combination.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>