[ST40] Rationalize SYSCONFs setup for ethernet on STx7108/STxH205/STxH415.
This commit rationalizes the way that "system configurations" registers
are initialized, specifically for ethernet pio/pad management, for the
following 3 STMicroelectronics' SoCs: STx7108, STxH205, and STxH415.

Initialization of the system configuration registers for the ethernet
PIOs for the aforementioned 3 SoCs was very similar, but there was no
real code re-use, despite the commonality.  This commit is essential a
re-factoring exercise to move the common structure+macro definitions, and
the common initialization code from SoC-specific files into arch-specific
files. This should both minimize maintenance costs, and provide a good
base for adding any new SoCs, which are all expected to re-use this code.

Now functions can simply call stm_configure_sysconfs() to configure
an entire array of SYSCONF registers. This is more in keeping with the
paradigm used in STLinux.

Only minimal testing: just built for one board of each of the 3 SoCs,
and executed on a single board with a STx7108 SoC.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>