[ST40] Rationalize PIO PAD setup for ethernet on STx7108/STxH205/STxH415.
This commit rationalizes the way that PIO "pad configuration" registers
are initialized, specifically for ethernet management, for the following
3 STMicroelectronics' SoCs: STx7108, STxH205, and STxH415.

Initialization of the pad configuration registers for the ethernet
PIOs for the aforementioned 3 SoCs was very similar, but there was no
real code re-use, despite the commonality.  This commit is essential a
re-factoring exercise to move the common structure+macro definitions, and
the common code from SoC-specific files into an arch-specific header file
(include/asm-st40/pad.h). This should both minimize maintenance costs,
and provide a good base for adding any new SoCs, which are all expected
to re-use this code.

Also performed some janitorial tidy up, to remove some macros which
added little to the comprehension of the code.

Only minimal testing: just built for one board of each of the 3 SoCs,
and executed on a single board with a STx7108 SoC.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>