Export "TARGET" to be used by "make" system.
The "mkconfig" script is passed several arguments. The first (after any
options) is the "Target", the 2nd is the "Architecture", the 3rd is the
"CPU", and the 4th is the "Board". Prior to this update, only the 2nd,
3rd, and 4th parameters were exported to config.mk as (the variables)
"ARCH", "CPU", and "BOARD", respectively. This commit essentially also
now passes the 1st parameter as (the variable) "TARGET" to the same
configuration file.

This allows other Makefiles to use the $(TARGET) variable instead of
the $(BOARD) variable, if desired.

Consider the following abridged example, where we have a board (B2057),
and a variant of it, that is so similar that we have chosen to share
the same board name and structure (only the 1st parameter is different).

b2057se_config :
	...
	$(MKCONFIG) -a b2057 st40 st40 b2057 st stxh205
	...

b2067se_config :
	...
	$(MKCONFIG) -a b2067 st40 st40 b2057 st stxh205
	...

Both will result in "BOARD=b2057". However, the first will now have
"TARGET=b2057", and the second will now have "TARGET=b2067". We can then,
in board/st/b2057/Makefile, have:
			SOBJS-y	= init-$(TARGET).o
instead of:
			SOBJS-y	= init-$(BOARD).o

Which will select the target-specific file (not the board-specific
file), since both files exist in the *same* board directory. There is
now no confusion with their co-existence, and no need for conditional
compilation.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>