[ST40] Pass the "-n" option to MKCONFIG for ST40 boards.
This is a minor (benign) change that passes the "-n" option to the
MKCONFIG script, so that the board name that we are configuring for,
(as specified in the make target) is printed in full (modulo the
"_config" suffix).

Prior to this change, running "make mb680se_config" would generate:

	host% make mb680se_config
	Configuring for mb680 board...

Now, it will generate:

	host% make mb680se_config
	Configuring for mb680se board...

This just helps when running ./MAKEALL to identify the precise target
that is actually being tested. There is no change in functionality.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>