[ST40] Add initial support for the B2067 Board (STxH238)
This commit adds initial support for the B2067A board from
STMicroelectronics, which is also known as the STiH238-HDK board, which
has a STiH238 SoC device (a derivative of the STiH205/STiH207).

This board is very similar to a previously supported board (the B2057).
It is so similar that it was decided to reuse the existing board/st/b2057
directory and the files therein for it, and not create a separate
directory (board/st/b2067), as one would normally do to support a
new board.

Only minimal testing of this support has been done, and only on a single
B2067A board (STiH238-HDKA003 E013).

Testing has included support for: Networking (with on-die IC+ IP101G PHY);
SPI serial FLASH; and multi-bit ECC NAND (MT29F4G08ABADAH4) using the
BCH controller in 18-bit ECC mode.  In addition, both boot-from-SPI, and
boot-from-NAND has been tested, and the board was able to boot from both
FLASH types with no problems.  NAND was only tested with 18-bit ECC on
the BCH controller, as the fitted NAND device is not capable of supporting
the ECC needs for 30-bit ECC, and using a weaker ECC is not recommended!

Note: Please ensure you read the comments near the beginning of the
"include/configs/b2067.h" file for details about setting up the mode pins
correctly. The ones in this file have been empirically verified for the
specific NAND device on the board (MT29F4G08ABADAH4), which is a 512 MiB
NAND, with a page size of 2048+64 bytes and a sector size of 128 KiB,
requiring 5 address cycles. If you are using a different NAND device,
then you may need to modify the mode-pins to boot from it - check the
SoC's data-sheet for details.

Note: It is recommended that the romgen utility is re-run and the file
"board/st/b2057/b2067-stxh205.romgen" is replaced, when building
an image to boot from FLASH. Please read the comments in the file
"board/st/b2057/init-b2067.S" for details on how to run romgen.  Finally,
please ensure the version of the TargetPack is "v06", or later.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>